Refactored so jwt module/service use jwt constants and interfaces.
