+++ src/jwt/jwt.constants.ts
@@ -1,0 +1,1 @@
+export const CONFIG_OPTIONS = 'CONFIG_OPTIONS';
