+++ src/jwt/jwt.service.ts
@@ -1,11 +1,12 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { JwtModuleOptions } from './interfaces/jwt-module-options.interface';
+import { JwtModuleOptions } from './jwt.interfaces';
+import { CONFIG_OPTIONS } from './jwt.constants';
 
 @Injectable()
 export class JwtService {
-  constructor(@Inject('BANANAS') private readonly options: JwtModuleOptions) {
-    console.log(options);
-  }
+  constructor(
+    @Inject(CONFIG_OPTIONS) private readonly options: JwtModuleOptions,
+  ) {}
   hello() {
     console.log('hello');
   }
